+++ lib/smart_core/container/arbitrary_lock.rb
@@ -2,7 +2,7 @@
 
 # @api private
 # @since 0.1.0
-class SmartCore::Container::ArbitaryLock
+class SmartCore::Container::ArbitraryLock
   # @return [void]
   #
   # @api private