+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.8.0] - 2020-07-08
+### Added
+- An ability to observe dependency re-registrations:
+  - `#observe(path, &observer) # => observer object` - listen specific dependency path;
+  - `#unobserve(observer)` - unsubscribe concrete observer object;
+  - `#clear_observers(path = nil)` - unsubscribe specific listenr or all listeners (`nil` parameter);
+
 ## [0.7.0] - 2020-06-20
 ### Added
 - `SmartCore::Container.define {}` - an ability to avoid explicit class definition that allows